Moved api configurations to `src/config/api.js`
